# backend/src/main/resources/api.yaml
openapi: 3.0.1
info:
  title: Conquery API
  description: Api to use Conquery.
  version: 2.0.0
...
paths:
  /datasets/{dataset}/concepts/{concept}:
    delete:
      tags:
        - admin
      operationId: removeConcept
      parameters:
        - name: dataset
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Dataset'
        - name: concept
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/ConceptObject'
      responses:
        default:
          description: default response
          content:
            application/json: {}
            application/x-jackson-smile: {}
...
  /datasets/{dataset}/tables/{table}/imports/{importId}:
    delete:
      tags:
        - admin
      operationId: deleteImport
      parameters:
        - name: dataset
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Dataset'
        - name: table
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Table'
        - name: importId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        default:
          description: default response
          content:
            application/json: {}
            application/x-jackson-smile: {}
  /datasets/{dataset}/tables/{table}/imports:
    get:
      tags:
        - admin
      operationId: listImports
      parameters:
        - name: dataset
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Dataset'
        - name: table
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Table'
      responses:
        default:
          description: default response
          content:
            application/json; charset=utf-8:
              schema:
                type: array
                items:
                  type: string
...
components:
  schemas:
    Dataset:
      required:
        - label
      type: object
      properties:
        name:
          type: string
        label:
          type: string
    ConceptObject:
      required:
        - label
        - type
      type: object
      properties:
        name:
          type: string
        label:
          type: string
        description:
          type: string
        additionalInfos:
          type: array
          items:
            $ref: '#/components/schemas/KeyValue'
        hidden:
          type: boolean
        connectors:
          type: array
          items:
            type: object
        dataset:
          $ref: '#/components/schemas/Dataset'
        selects:
          type: array
          items:
            $ref: '#/components/schemas/Select'
        defaultSelects:
          type: array
          items:
            $ref: '#/components/schemas/Select'
        type:
          type: string
      discriminator:
        propertyName: type
    KeyValue:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
    Select:
      required:
        - label
        - type
      type: object
      properties:
        name:
          type: string
        label:
          type: string
        holder:
          $ref: '#/components/schemas/SelectHolderObject'
        description:
          type: string
        default:
          type: boolean
        type:
          type: string
      discriminator:
        propertyName: type
    SelectHolderObject:
      type: object
      properties:
        selects:
          type: array
          items:
            type: object
    SecondaryIdDescription:
      required:
        - label
      type: object
      properties:
        name:
          type: string
        label:
          type: string
        dataset:
          $ref: '#/components/schemas/Dataset'
        description:
          type: string
    Column:
      required:
        - label
        - table
        - type
      type: object
      properties:
        name:
          type: string
        label:
          type: string
        table:
          $ref: '#/components/schemas/Table'
        type:
          type: string
          enum:
            - STRING
            - INTEGER
            - BOOLEAN
            - REAL
            - DECIMAL
            - MONEY
            - DATE
            - DATE_RANGE
        position:
          type: integer
          format: int32
        sharedDictionary:
          type: string
        secondaryId:
          $ref: '#/components/schemas/SecondaryIdDescription'
    Table:
      required:
        - columns
        - label
      type: object
      properties:
        name:
          type: string
        label:
          type: string
        dataset:
          $ref: '#/components/schemas/Dataset'
        columns:
          type: array
          items:
            $ref: '#/components/schemas/Column'
...
    Id:
      pattern: '[a-z0-9](\.[a-z0-9]+)*'
      type: object
      format: string
...